Breadcrumbs/Mongo integration: Minor test and efficieny improvements

+++ features/fixtures/rails5/app/app/models/mongo_model.rb
@@ -2,5 +2,5 @@ class MongoModel
   include Mongoid::Document
 
   field :string_field, type: String
-  field :numeric_field, type: Numeric
+  field :numeric_field, type: Integer
 end
